requirements.txt
django-health-check==1.0.2
tox